app/requirements.txt
Flask
mysql-connector
python-dateutil
requests
